Moved build arguments into environment variables.
Secrets were given to the container as build arguments (as opposed to
environment variables) because that allowed more control in terms of how
to handle missing secrets. One of the problems was that if
'docker-compose.yml' specified mandatory environment variables in
multiple services, it would require all of those environment variables
even if you tried to start only one service. But the downside of using
build arguments for this is that secrets are now baked into the images.

It is better to follow best practices and not include secrets into
images.