refactor: Combines flight record and data into one resource.
Flight record and flight data were two separate resources with
one-to-one relationship. Flight record contained all of the metadata,
and flight data all of the data points. This allowed API structure where
client could retrieve list of all flights containing just the metadata.
And the associated flight data, which is much larger and slower to
fetch, could be fetched later on if needed.

But this created a problem because now these two resources had to be
also created separately. And in the (although unlikely) scenario that
flight record creation succeeded and flight data creation failed, the
system would end up in a problematic state. And this state would be
rather difficult to communicate to the user without having to go into
the details of the system. And there would have to be also a way to fix
this situation.

The solution is to combine these two into one resource, because these
have one-to-one relationship and have to be created at the same time.
And when client queries all of the flights, only the metadata is
returned. And if the client needs the entire resource, it can be
accessed with the resource id.